Split image viewer, dialog, and panel into separate components
