+++ frontend/app/common/image-viewer/image-viewer.tsx
@@ -1,123 +1,16 @@
-import {
-  ElementType,
-  ForwardedRef,
-  Fragment,
-  SetStateAction,
-  forwardRef,
-  useCallback,
-  useEffect,
-  useRef,
-  useState,
-} from "react";
 import { Image } from "../image";
-import { Dialog, Transition } from "@headlessui/react";
 import { cn } from "../cn";
-import {
-  ArrowLeftIcon,
-  ArrowRightIcon,
-  ArrowsPointingOutIcon,
-} from "@heroicons/react/20/solid";
-import { useTranslation } from "react-i18next";
+import { ImageViewerImage } from "./types";
+import { useState } from "react";
+import { ImageViewerDialog } from "./image-viewer-dialog";
 
 export type ImageViewerProps = {
-  images: {
-    src: string;
-    alt: string;
-    caption?: string;
-  }[];
+  images: ImageViewerImage[];
 };
-
-function isInFullscreen() {
-  return !!document.fullscreenElement;
-}
 
 export function ImageViewer({ images }: ImageViewerProps) {
   const [isOpen, setIsOpen] = useState(false);
   const [currentImageIndex, setCurrentImageIndex] = useState(0);
-  const [isZoomed, setIsZoomed] = useState(false);
-  const zoomButtonRef = useRef<HTMLButtonElement>(null);
-  const [isUserActive, setIsUserActive] = useState(true);
-  const [isFullscreen, setIsFullscreen] = useState(isInFullscreen());
-  const activityTimeoutRef = useRef(0);
-
-  // sync fullscreen state with browser
-  useEffect(() => {
-    function handleFullscreenChange() {
-      setIsFullscreen(isInFullscreen());
-    }
-
-    document.documentElement.addEventListener(
-      "fullscreenchange",
-      handleFullscreenChange,
-    );
-
-    return () =>
-      document.documentElement.removeEventListener(
-        "fullscreenchange",
-        handleFullscreenChange,
-      );
-  }, []);
-
-  useEffect(() => {
-    if (!isFullscreen) setIsZoomed(false);
-  }, [isFullscreen]);
-
-  useEffect(() => {
-    function handleActivity() {
-      setIsUserActive(true);
-      resetTimer();
-    }
-
-    function resetTimer() {
-      if (activityTimeoutRef.current) clearTimeout(activityTimeoutRef.current);
-      activityTimeoutRef.current = window.setTimeout(() => {
-        setIsUserActive(false);
-      }, 3000);
-    }
-
-    document.addEventListener("mousemove", handleActivity);
-    document.addEventListener("keydown", handleActivity);
-    return () => {
-      document.removeEventListener("mousemove", handleActivity);
-      document.removeEventListener("keydown", handleActivity);
-    };
-  }, []);
-
-  const goToPreviousImage = useCallback(
-    function goToPreviousImage() {
-      goToImage((v) => (v > 0 ? v - 1 : images.length - 1));
-    },
-    [images.length],
-  );
-
-  const goToNextImage = useCallback(
-    function goToNextImage() {
-      goToImage((v) => (v < images.length - 1 ? v + 1 : 0));
-    },
-    [images.length],
-  );
-
-  useEffect(() => {
-    function handleKeyDown(e: KeyboardEvent) {
-      if (e.key === "ArrowLeft") {
-        goToPreviousImage();
-      } else if (e.key === "ArrowRight") {
-        goToNextImage();
-      } else if (isZoomed && e.key === "Escape") {
-        e.stopPropagation();
-        setIsZoomed(false);
-      }
-    }
-
-    document.addEventListener("keydown", handleKeyDown);
-    return () => document.removeEventListener("keydown", handleKeyDown);
-  }, [goToNextImage, goToPreviousImage, isZoomed]);
-
-  function goToImage(value: SetStateAction<number>) {
-    setIsZoomed(false);
-    setCurrentImageIndex(value);
-  }
-
   return (
     <>
       <div className="grid w-[40rem] grid-cols-5 grid-rows-[auto,auto] gap-2">
@@ -160,178 +53,13 @@ export function ImageViewer({ images }: ImageViewerProps) {
           </button>
         ))}
       </div>
-
-      <Transition show={isOpen} as={Fragment}>
-        <Dialog
-          onClose={() => setIsOpen(false)}
-          className="relative z-50"
-          initialFocus={zoomButtonRef}
-        >
-          <Transition.Child
-            as={Fragment}
-            enter="ease-out duration-300"
-            enterFrom="opacity-0"
-            enterTo="opacity-100"
-            leave="ease-in duration-200"
-            leaveFrom="opacity-100"
-            leaveTo="opacity-0"
-          >
-            <div className="fixed inset-0 bg-black" aria-hidden="true" />
-          </Transition.Child>
-          <Dialog.Panel
-            className={cn(
-              "m-auto",
-              isZoomed
-                ? "fixed inset-0 overflow-auto"
-                : cn(
-                    "fixed inset-0 w-max",
-                    !isFullscreen &&
-                      "max-h-[calc(100%-6rem)] max-w-[calc(100%-6rem)]",
-                  ),
-            )}
-          >
-            {isUserActive && <PreviousButton onClick={goToPreviousImage} />}
-            <button
-              ref={zoomButtonRef}
-              className={cn(
-                "block h-full w-full focus:outline-none",
-                isZoomed ? "cursor-all-scroll" : "cursor-zoom-in",
-              )}
-              onClick={() => setIsZoomed(!isZoomed)}
-            >
-              <Transition.Child
-                as={Fragment}
-                enter="ease-out duration-300"
-                enterFrom="-translate-x-96 -translate-y-24 scale-50 opacity-0"
-                enterTo="translate-x-0 translate-y-0 scale-100 opacity-100"
-                leave="ease-in duration-200"
-                leaveFrom="translate-x-0 translate-y-0 scale-100 opacity-100"
-                leaveTo="-translate-x-96 -translate-y-24 scale-50 opacity-0"
-              >
-                <Image
-                  src={images[currentImageIndex].src}
-                  alt={images[currentImageIndex].alt}
-                  className={cn(
-                    isZoomed ? "h-max w-max scale-110" : "h-full w-full",
-                  )}
-                  {...(isZoomed
-                    ? undefined
-                    : {
-                        transformation: {
-                          height: document.body.clientHeight * 2,
-                        },
-                      })}
-                />
-              </Transition.Child>
-            </button>
-            {isUserActive && <NextButton onClick={goToNextImage} />}
-            {isUserActive && (
-              <div className="fixed inset-0 flex h-12 w-full items-center justify-between gap-4 bg-black/60 px-4 text-sm text-neutral-300 shadow-lg">
-                <div>
-                  {currentImageIndex + 1} / {images.length}
-                </div>
-                <div>{images[currentImageIndex].caption}</div>
-                <div>{!isFullscreen && <FullscreenButton />}</div>
-              </div>
-            )}
-          </Dialog.Panel>
-        </Dialog>
-      </Transition>
+      <ImageViewerDialog
+        key={currentImageIndex}
+        images={images}
+        isOpen={isOpen}
+        onDismiss={() => setIsOpen(false)}
+        defaultImageIndex={currentImageIndex}
+      />
     </>
   );
 }
-
-function PreviousButton({ onClick }: Pick<PreviousNextButtonProps, "onClick">) {
-  const { t } = useTranslation();
-  return (
-    <PreviousNextButton
-      className={cn("left-0 top-1/2 -translate-y-1/2")}
-      icon={ArrowLeftIcon}
-      onClick={onClick}
-      title={t("imageViewer.previous")}
-    />
-  );
-}
-
-function NextButton({ onClick }: Pick<PreviousNextButtonProps, "onClick">) {
-  const { t } = useTranslation();
-  return (
-    <PreviousNextButton
-      className={cn("right-0 top-1/2 -translate-y-1/2")}
-      icon={ArrowRightIcon}
-      onClick={onClick}
-      title={t("imageViewer.next")}
-    />
-  );
-}
-
-function FullscreenButton() {
-  const { t } = useTranslation();
-  return (
-    <IconButton
-      icon={ArrowsPointingOutIcon}
-      onClick={() => document.documentElement.requestFullscreen()}
-      title={t("imageViewer.fullscreen")}
-    />
-  );
-}
-
-type PreviousNextButtonProps = {
-  className?: string;
-  icon: ElementType;
-  onClick: () => void;
-  title: string;
-};
-
-function PreviousNextButton({
-  className,
-  onClick,
-  icon,
-  title,
-}: PreviousNextButtonProps) {
-  return (
-    <Transition.Child
-      as={Fragment}
-      enter="delay-200"
-      enterFrom="invisible"
-      enterTo="visible"
-    >
-      <IconButton
-        icon={icon}
-        onClick={onClick}
-        title={title}
-        className={cn(
-          "fixed z-10 m-4 rounded-full bg-black/60 shadow-md",
-          className,
-        )}
-      />
-    </Transition.Child>
-  );
-}
-
-type IconButtonProps = {
-  className?: string;
-  icon: ElementType;
-  onClick: () => void;
-  title: string;
-};
-
-const IconButton = forwardRef(function IconButton(
-  { className, icon: Icon, onClick, title }: IconButtonProps,
-  ref: ForwardedRef<HTMLButtonElement>,
-) {
-  return (
-    <button
-      className={cn(
-        "p-2 text-neutral-300 hover:text-white focus:text-white",
-        className,
-      )}
-      onClick={onClick}
-      title={title}
-      ref={ref}
-    >
-      <span className="sr-only">{title}</span>
-      <Icon className="h-5 w-5" />
-    </button>
-  );
-});
