+++ frontend/app/common/image-viewer/types.ts
@@ -1,0 +1,5 @@
+export type ImageViewerImage = {
+  src: string;
+  alt: string;
+  caption?: string;
+};
